+++ kamon-cassandra-client/src/main/scala/com/datastax/driver/core/ConnectionPoolAdvices.scala
@@ -5,7 +5,7 @@ import java.util.concurrent.atomic.AtomicInteger
 import com.google.common.util.concurrent.{FutureCallback, ListenableFuture}
 import kamon.Kamon
 import kamon.instrumentation.cassandra.metrics.{HasPoolMetrics, PoolMetrics}
-import kamon.instrumentation.cassandra.Cassandra
+import kamon.instrumentation.cassandra.CassandraInstrumentation
 import kamon.instrumentation.cassandra.metrics.PoolMetrics.PoolInstruments
 import kanela.agent.libs.net.bytebuddy.asm.Advice
 
@@ -13,7 +13,7 @@ import kanela.agent.libs.net.bytebuddy.asm.Advice
 object PoolConstructorAdvice {
   @Advice.OnMethodExit
   def onConstructed(@Advice.This poolWithMetrics: HasPoolMetrics, @Advice.FieldValue("host") host: Host): Unit = {
-    poolWithMetrics.setMetrics(new PoolInstruments(Cassandra.targetFromHost(host)))
+    poolWithMetrics.setMetrics(new PoolInstruments(CassandraInstrumentation.targetFromHost(host)))
   }
 }
